requirements.txt
llama-index
python-dotenv
html2text
llama-index-llms-gemini
llama-index-llms-google-genai
llama-index-readers-web
llama_index-embeddings-gemini